Rename converter.py to processing.py and some minor changes
